+++ src/BlazorDB/IStorageContext.cs
@@ -1,12 +1,7 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-
-namespace BlazorDB
+﻿namespace BlazorDB
 {
     public interface IStorageContext
     {
-        void SaveChanges();
+        int SaveChanges();
     }
 }
