+++ src/BlazorDB/ServiceCollectionExtensions.cs
@@ -1,4 +1,5 @@
-﻿using Microsoft.AspNetCore.Blazor;
+﻿using BlazorDB.Storage;
+using Microsoft.AspNetCore.Blazor;
 using Microsoft.Extensions.DependencyInjection;
 using System;
 using System.Collections.Generic;
