+++ src/BlazorDB/Storage/Metadata.cs
@@ -1,0 +1,12 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Threading.Tasks;
+
+namespace BlazorDB.Storage
+{
+    internal class Metadata
+    {
+
+    }
+}
